+++ superduperdb/base/variables.py
@@ -23,6 +23,8 @@ def _replace_variables(x, **kwargs):
             _replace_variables(k, **kwargs): _replace_variables(v, **kwargs)
             for k, v in x.items()
         }
+    if isinstance(x, str) and re.match(r'^<var:(.*?)>$', x) is not None:
+        return kwargs.get(x[5:-1], x)
     if isinstance(x, str):
         variables = re.findall(r'<var:(.*?)>', x)
         variables = list(map(lambda v: v.strip(), variables))
@@ -31,7 +33,8 @@ def _replace_variables(x, **kwargs):
                 if isinstance(v, str):
                     x = x.replace(f'<var:{k}>', v)
                 else:
-                    x = v
+                    x = re.sub('[<>:]', '-', x)
+                    x = re.sub('[-]+', '-', x)
         return x
     if isinstance(x, (list, tuple)):
         return [_replace_variables(v, **kwargs) for v in x]