+++ CHANGELOG.md
@@ -15,7 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 #### New Features & Functionality
 
-- 
+- QueryTemplate component
 
 #### Bug Fixes
 
@@ -42,7 +42,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Mask special character keys in mongodb queries
 - Fix listener cleanup after removal
 - Don't require `@dc.dataclass` or `@merge_docstrings` decorator
-- Make output of `Document.encode()` a bit more minimalistic
+- Make output of `Document.encode()` more minimalistic
 - Increment minimum supported ibis version to 9.0.0
 - Make database connections reconnection on token expiry
 - Prototype the cron job services