+++ CHANGELOG.md
@@ -83,6 +83,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add `Application` and `Template` support to build reusable apps
 - Add pretty-print to `Component.info`
 - `Model` 
+- 'Add pluggable compute backend via config'
 
 #### Bug Fixes