+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 #### Bug Fixes
 
 - Fix templates
+- Fixed the bug when exporting the component.
 
 ## [0.2.0](https://github.com/SuperDuperDB/superduperdb/compare/0.1.3...0.2.0])    (2024-Jun-21)