+++ CHANGELOG.md
@@ -20,7 +20,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 #### Bug Fixes
 
 - Fix templates
-- Fixed the bug when exporting the component.
+- Fix the issue of the filter in select not working in the listener.
+- Fix exports
+- Fix model query
 
 ## [0.2.0](https://github.com/SuperDuperDB/superduperdb/compare/0.1.3...0.2.0])    (2024-Jun-21)